requirements.txt
appnope==0.1.4
asttokens==2.4.1
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
executing==2.0.1
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
huggingface-hub==0.24.6
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.2
packaging==24.1
pandas==2.1.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyaes==1.6.1
pyasn1==0.6.0
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.2
pyzmq==26.0.3
regex==2024.7.24
requests==2.31.0
rsa==4.9
safetensors==0.4.4
scipy==1.14.0
seaborn==0.13.2
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.23
stack-data==0.6.3
sympy==1.13.2
Telethon==1.36.0
tokenizers==0.19.1
torch==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
widgetsnbextension==4.0.11